+++ tortoise/__init__.py
@@ -708,4 +708,4 @@ def run_async(coro: Coroutine) -> None:
         loop.run_until_complete(Tortoise.close_connections())
 
 
-__version__ = "0.17.1"
+__version__ = "0.17.2"