+++ tortoise/contrib/mysql/indexes.py
@@ -1,0 +1,5 @@
+from tortoise.indexes import Index
+
+
+class FullTextIndex(Index):
+    INDEX_CREATE_TEMPLATE = "CREATE FULLTEXT INDEX `{index_name}` ON `{table_name}` ({fields});"